Factor out assert_post to simplify load_demo_data

+++ tools/load_demo_data.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import shutil
 import pandas as pd
 import signal
+import requests
 
 from baselayer.app.env import load_env
 from baselayer.app.model_util import status, create_tables, drop_tables
@@ -52,9 +53,26 @@ if __name__ == "__main__":
             DBSession().add(TornadoStorage.user.create_social_auth(u, u.username,
                                                                    'google-oauth2'))
 
+    with status("Creating token"):
+        token = create_token(['Manage groups', 'Manage sources', 'Upload data',
+                              'Comment', 'Manage users'],
+                             super_admin_user.id,
+                             'load_demo_data token')
+
+    def assert_post(endpoint, data):
+        response_status, data = api('POST', endpoint, data, token)
+        if not response_status == 200 and data['status'] == 'success':
+            raise RuntimeError(f'API call to {endpoint} failed with status {status}: {data["message"]}')
+        return data
+
     with status("Launching web app & executing API calls"):
-        web_client = subprocess.Popen(['make', 'run'],
-                                      cwd=basedir, preexec_fn=os.setsid)
+        try:
+            response_status, data = api('GET', 'sysinfo', token=token)
+            app_already_running = True
+        except requests.ConnectionError:
+            app_already_running = False
+            web_client = subprocess.Popen(['make', 'run'],
+                                          cwd=basedir, preexec_fn=os.setsid)
 
         server_url = f"http://localhost:{cfg['ports.app']}"
         print()
@@ -63,60 +81,39 @@ if __name__ == "__main__":
         try:
             verify_server_availability(server_url)
             print('App running - continuing with API calls')
-            with status("Creating token"):
-                token = create_token(['Manage groups', 'Manage sources', 'Upload data',
-                                      'Comment', 'Manage users'],
-                                     super_admin_user.id,
-                                     'load_demo_data token')
 
             with status("Creating dummy group & adding users"):
-                response_status, data = api(
-                    'POST',
-                    'groups',
-                    data={'name': 'Stream A',
-                          'group_admins': [super_admin_user.username,
-                                           group_admin_user.username]},
-                    token=token)
-                assert response_status == 200
-                assert data['status'] == 'success'
+                data = assert_post('groups',
+                                   data={'name': 'Stream A',
+                                         'group_admins': [super_admin_user.username,
+                                                          group_admin_user.username]})
                 group_id = data['data']['id']
 
                 for u in [view_only_user, full_user]:
-                    response_status, data = api(
-                        'POST',
+                    data = assert_post(
                         f'groups/{group_id}/users/{u.username}',
-                        data={'admin': False},
-                        token=token)
-                    assert response_status == 200
-                    assert data['status'] == 'success'
+                        data={'admin': False})
 
             with status("Creating dummy instruments"):
-                response_status, data = api('POST', 'telescope',
-                                            data={'name': 'Palomar 1.5m',
-                                                  'nickname': 'P60',
-                                                  'lat': 33.3633675,
-                                                  'lon': -116.8361345,
-                                                  'elevation': 1870,
-                                                  'diameter': 1.5,
-                                                  'group_ids': [group_id]
-                                            },
-                                            token=token)
-                assert response_status == 200
-                assert data['status'] == 'success'
+                data = assert_post(
+                    'telescope',
+                    data={'name': 'Palomar 1.5m',
+                          'nickname': 'P60',
+                          'lat': 33.3633675,
+                          'lon': -116.8361345,
+                          'elevation': 1870,
+                          'diameter': 1.5,
+                          'group_ids': [group_id]})
                 telescope1_id = data['data']['id']
 
-                response_status, data = api('POST', 'instrument',
-                                            data={'name': 'P60 Camera',
-                                                  'type': 'phot',
-                                                  'band': 'optical',
-                                                  'telescope_id': telescope1_id
-                                            },
-                                            token=token)
-                assert response_status == 200
-                assert data['status'] == 'success'
+                data = assert_post('instrument',
+                                   data={'name': 'P60 Camera',
+                                         'type': 'phot',
+                                         'band': 'optical',
+                                         'telescope_id': telescope1_id})
                 instrument1_id = data['data']['id']
 
-                response_status, data = api('POST', 'telescope',
+                data = assert_post('telescope',
                                             data={'name': 'Nordic Optical Telescope',
                                                   'nickname': 'NOT',
                                                   'lat': 28.75,
@@ -124,21 +121,15 @@ if __name__ == "__main__":
                                                   'elevation': 1870,
                                                   'diameter': 2.56,
                                                   'group_ids': [group_id]
-                                            },
-                                            token=token)
-                assert response_status == 200
-                assert data['status'] == 'success'
+                                            })
                 telescope2_id = data['data']['id']
 
-                response_status, data = api('POST', 'instrument',
-                                            data={'name': 'ALFOSC',
-                                                  'type': 'both',
-                                                  'band': 'optical',
-                                                  'telescope_id': telescope2_id
-                                            },
-                                            token=token)
-                assert response_status == 200
-                assert data['status'] == 'success'
+                data = assert_post('instrument',
+                                   data={'name': 'ALFOSC',
+                                         'type': 'both',
+                                         'band': 'optical',
+                                         'telescope_id': telescope2_id
+                                   })
 
             with status("Creating dummy sources"):
                 SOURCES = [{'id': '14gqr', 'ra': 353.36647, 'dec': 33.646149,
@@ -153,69 +144,55 @@ if __name__ == "__main__":
                 for source_info in SOURCES:
                     comments = source_info.pop('comments')
 
-                    response_status, data = api('POST', 'sources',
-                                                data=source_info,
-                                                token=token)
-                    assert response_status == 200
+                    data = assert_post('sources', data=source_info)
                     assert data['data']['id'] == source_info['id']
 
                     for comment in comments:
-                        response_status, data = api('POST', 'comment',
-                                                    data={'source_id': source_info['id'],
-                                                          'text': comment},
-                                                    token=token)
-                        assert response_status == 200
+                        data = assert_post('comment',
+                                           data={'source_id': source_info['id'],
+                                                 'text': comment})
 
                     phot_file = os.path.join(os.path.dirname(os.path.dirname(__file__)),
                                              'skyportal', 'tests', 'data',
                                              'phot.csv')
                     phot_data = pd.read_csv(phot_file)
 
-                    response_status, data = api('POST', 'photometry',
-                                                data={'source_id': source_info['id'],
-                                                      'time_format': 'iso',
-                                                      'time_scale': 'utc',
-                                                      'instrument_id': instrument1_id,
-                                                      'observed_at': phot_data.observed_at.tolist(),
-                                                      'mag': phot_data.mag.tolist(),
-                                                      'e_mag': phot_data.e_mag.tolist(),
-                                                      'lim_mag': phot_data.lim_mag.tolist(),
-                                                      'filter': phot_data['filter'].tolist()
-                                                },
-                                                token=token)
-                    assert response_status == 200
-                    assert data['status'] == 'success'
+                    data = assert_post('photometry',
+                                       data={'source_id': source_info['id'],
+                                             'time_format': 'iso',
+                                             'time_scale': 'utc',
+                                             'instrument_id': instrument1_id,
+                                             'observed_at': phot_data.observed_at.tolist(),
+                                             'mag': phot_data.mag.tolist(),
+                                             'e_mag': phot_data.e_mag.tolist(),
+                                             'lim_mag': phot_data.lim_mag.tolist(),
+                                             'filter': phot_data['filter'].tolist()
+                                       })
 
                     spec_file = os.path.join(os.path.dirname(os.path.dirname(__file__)),
                                              'skyportal', 'tests', 'data',
                                              'spec.csv')
                     spec_data = pd.read_csv(spec_file)
                     for i, df in spec_data.groupby('instrument_id'):
-                        response_status, data = api(
-                            'POST', 'spectrum',
+                        data = assert_post(
+                            'spectrum',
                             data={'source_id': source_info['id'],
                                   'observed_at': str(datetime.datetime(2014, 10, 24)),
                                   'instrument_id': 1,
                                   'wavelengths': df.wavelength.tolist(),
                                   'fluxes': df.flux.tolist()
-                            },
-                            token=token)
-                        assert response_status == 200
-                        assert data['status'] == 'success'
+                            })
 
                     for ttype in ['new', 'ref', 'sub']:
                         fname = f'{source_info["id"]}_{ttype}.png'
                         fpath = basedir/f'skyportal/tests/data/{fname}'
                         thumbnail_data = base64.b64encode(
                             open(os.path.abspath(fpath), 'rb').read())
-                        response_status, data = api('POST', 'thumbnail',
-                                                    data={'source_id': source_info['id'],
-                                                          'data': thumbnail_data,
-                                                          'ttype': ttype
-                                                    },
-                                                    token=token)
-                        assert response_status == 200
-                        assert data['status'] == 'success'
+                        data = assert_post('thumbnail',
+                                           data={'source_id': source_info['id'],
+                                                 'data': thumbnail_data,
+                                                 'ttype': ttype
+                                           })
 
                     source = Source.query.get(source_info['id'])
                     source.add_linked_thumbnails()
@@ -223,5 +200,6 @@ if __name__ == "__main__":
             print(e)
             raise
         finally:
-            print('Terminating web app')
-            os.killpg(os.getpgid(web_client.pid), signal.SIGTERM)
+            if not app_already_running:
+                print('Terminating web app')
+                os.killpg(os.getpgid(web_client.pid), signal.SIGTERM)
